+++ fe/src/api/stock/queries/useStockSearchQuery.ts
